requirements.txt
# Core RAG
langchain
sentence-transformers
faiss-cpu

# Document parsing
pypdf

# Utilities
python-dotenv
tqdm
